Initial check in of generic action interface node for industrial robot client package.  The node bridges the gap between the arm navigation/moveit libraries and the ROS-Industrial driver spec.
The getJointNames from parameter function was also modified to remove the number of joints argument.  This argument was used to initialize a default set of joints if the parameter was not populated.  This function is now used across several nodes, requiring the number of joints argument to be syncronized across all nodes.  This would likely be hard to maintain and could cause headaches when extending the base client classes.  For this reason, the number of joint argument was removed.

git-svn-id: https://swri-ros-pkg.googlecode.com/svn/trunk@1066 076cdf4d-ed99-c8c1-5d01-9bc0d27e81bd