driver: remove default for J23 coupling parameter.
Users should explicitly provide this on the command line, or use one
of the convenience launchfiles provided with the support packages.

It is of critical importance that this parameter is set to the
correct value, and should therefore not be supplied a default in
a (for end-users) file with low visibility.